OPEN_AI_AGENTS_SDK_NEW/14_basic_handoffs/pyproject.toml
[project]
name = "14-basic-handoffs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.9",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
